Class should be unique per PHP file
